[Kernel/Core] Redefined size_t exactly as GCC expects it.

+++ Kernel/Core/Types.h
@@ -12,7 +12,7 @@ using u16 = unsigned short int;
 using u32 = unsigned int;
 using u64 = unsigned long long int;
 
-using size_t = u32;
+using size_t = long unsigned int;
 using ptr_t  = u32;
 
 // floating point
